added eater
Eaters are good if you have lots of teaser or advertisement items which
you don't necessarily need on every resolution and can be missed.

They are called 'eater' because those elements 'eat' content items if
the browser width gets smaller.

Actually, this approach here is more of an 'spitter' than an 'eater',
because it's implemented mobile-first.

The eater plugin here requires two parameters, which are located in the
_variables file:

$eaterWidth, which is set to a standard grid with of 960px
$eaterItems, which is 4 by default